requirements.txt
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
readability-lxml>=0.8.1
dateparser>=1.1
python-slugify>=8.0
python-dateutil>=2.8
tqdm>=4.66
